requirements.txt
boto3>=1.28.0
chromadb>=0.4.22
jupyter>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
openai>=0.27.4
